+++ src/beatmap/schema/shared/types/environment.ts
@@ -48,6 +48,7 @@ export const EnvironmentName = [
    'GridEnvironment',
    'ColdplayEnvironment',
    'ProdigyEnvironment',
+   'HipHopMixtape2Environment',
    'GlassDesertEnvironment',
    'MultiplayerEnvironment',
 ] as const;
@@ -103,7 +104,8 @@ export type EnvironmentV3Name =
    | 'MetallicaEnvironment'
    | 'GridEnvironment'
    | 'ColdplayEnvironment'
-   | 'ProdigyEnvironment';
+   | 'ProdigyEnvironment'
+   | 'HipHopMixtape2Environment';
 
 /** List of available 360 environment in base game. */
 export type Environment360Name =