+++ src/extensions/heck/chroma/lightID.ts
@@ -1181,6 +1181,7 @@ export const LightIDList: {
    GridEnvironment: {},
    ColdplayEnvironment: {},
    ProdigyEnvironment: {},
+   HipHopMixtape2Environment: {},
    GlassDesertEnvironment: {
       '0': [1, 2, 3, 4, 5, 6],
       '1': [1, 2, 3, 4, 5, 6, 7, 8],